requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=1.8.0
moviepy>=1.0.3
imageio==2.31.1  # moviepy的依赖项
imageio-ffmpeg==0.4.8
python-multipart
beautifulsoup4>=4.9.3
tenacity==8.2.3
openai>=1.12.0  # 需要新版OpenAI SDK
dashscope>=1.14.0  # 阿里云官方SDK
Pillow>=10.0.0
imageio>=2.31.1
volcengine